{{cookiecutter.project_slug}}/requirements-dev.txt
-r requirements{% if cookiecutter.package_manager == "pip-tools" %}.in{%else %}.txt{%endif%}

# Code style / formatting
black==23.3.0
isort==5.12.0
flake8==5.0.4

#Testes
pytest==7.2.2
pytest-django==4.5.2
mock==5.1.0

# Better Dev Experience
ipython==8.16.1

# CORS (para quando o frontend está em domínio diferente)
{% if cookiecutter.docker_usage == "📦 use venv npm by default" %}
django-cors-headers==4.3.0
{% endif %}